working on admin page-corrected put and post
