+++ client/components/books.js
@@ -7,7 +7,8 @@ import {
   Image,
   Container,
   Button,
-  Divider
+  Divider,
+  Card
 } from 'semantic-ui-react'
 
 const mapStateToProps = state => {
@@ -63,7 +64,7 @@ class Books extends Component {
           {books.map(book => {
             return (
               <GridColumn width={4} className="container">
-                <Container>
+                <Card>
                   <Link to={`book/${book.id}`}>{book.name}</Link>
                   <p>by {book.author}</p>
                   {/* Book price is an integer in db, so we need to reformat it as a price */}
@@ -74,7 +75,7 @@ class Books extends Component {
                   </p>
                   <Image src={book.imgUrl} />
                   <Button icon="shop" />
-                </Container>
+                </Card>
                 <Divider hidden />
               </GridColumn>
             )
