+++ client/store/books.js
@@ -34,7 +34,7 @@ const requestAllBooksFromServer = () => {
 export const createBook = book => {
   return async dispatch => {
     try {
-      book.genres = book.genres.split(' ')
+      book.genres = book.genres.split(',')
       const {data} = await axios.post('/api/books', book)
       dispatch(addBook(data))
     } catch (err) {
@@ -46,7 +46,6 @@ export const createBook = book => {
 export const editBook = book => {
   return async dispatch => {
     try {
-      book.genres = book.genres.split(' ')
       const {data} = await axios.put('/api/books', book)
       dispatch(updateBook(data))
     } catch (err) {
@@ -85,9 +84,11 @@ export default function(state = initialState, action) {
       return {...state, list: [...state.list, action.book]}
     }
     case UPDATE_BOOK: {
-      const newBookArray = state.list.map(book => {
+      const newBookArray = state.list.slice().map(book => {
         if (book.id === action.book.id) {
-          book = action.book
+          return action.book
+        } else {
+          return book
         }
       })
       return {...state, list: newBookArray}
